chore: increase visibility of prod -> staging db sync (#32)
The sync that keeps the staging environment's DB up to date with
production was failing due to the staging app not having permissions to
access the prod DB. I fixed that, but am also changing it so that now if
the sync fails, it will cause the deployment to fail and I'll know about
it.

Eventually if the prod DB gets huge and this sync is slow, we can
revisit this and consider running the sync in the background again.