feat: users receive emails for new activity (#54)
Right now email is only used for registration and password reset. This
expands on that so that a user gets an email for all activity in the
app. That way they can respond promptly. Any kind of mobile app is a
long ways off so this is how people can stay on top of activity.

The migration is big because it is catching for a miss on tweaks to the
user account deletion feature that should have had a migration. I didn't
create a migration after implementing soft delete.